Fixed an error in the interrupt system
Due to improper handling of the timer interrupt status the core irq was set directly after the first response to it. The timer interrupt request needed to be cleared directly after propagation to intc and irq.
Also fixed a problem in the tracing when a double word instruction is called. Only the first word was logged.