Looked into the mmu setup
Renamed the functions for the mmu setup and for as far as I can see and understand it there is no real translation being done. It is just using section entries and signals that the dram is cacheable and the rest of the memory is not. The table is filled in a linear fashion