added mor menu functions
Renamed the display_ functions in the scope_functions files to scope_ to distinct them from the actual display functions. Added a display function for copying an icon to the screen buffer. Implemented the channel2 menu, the trigger menu, the move speed button and the main menu. The main menu only displays itself. The rest needs to be implemented yet.