Changed wait for touch release to a function
Needed wait on touch released when touch outside a menu was detected to avoid the menu being opened again when touch was on the menu open button. To reduce on the code a bit changed the wait while loops to function calls and added a wait for touch release function.