Started mass storage implementation
Started with a mass storage handler based on the sunxi usb handler. Copied the descriptors from the original scope code and started a crude implementation of the MSC CBW and CSW protocol. Setup a state machine for handling the different states. SCSI commands needed for initialization are also crudely  implemented. It works but needs checking on error conditions. Still to do is read and write.