Fixed the flicker of the display
The FPGA in the 1014D expects 0x78 written to command 0x38 to set the backlight PWM to fully on. Since the FPGA is not clocked yet due to the MS5351M clock synthesizer not being initialized yet, it can't generate a proper PWM signal.

Changed the FPGA backlight write to only send a byte value.